requirements.txt
Flask
keras
tensorflow
librosa
numpy
Pillow
